requirements.txt
Django==4.0.6
asgiref==3.5.2
astroid==2.5.1
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djongo==1.3.6
dnspython==2.2.1
httpsproxy-urllib2==1.0
idna==3.3
inflection==0.5.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mysql-connector-python==8.0.29
numpy==1.23.0
pandas==1.4.3
pip==22.2
protobuf==4.21.2
pylint==2.7.2
pymongo==4.2.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
setuptools==47.1.0
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
tzdata==2022.1
urllib2-file==0.2.1
urllib3==1.26.9